haystack: add toDateSpan, toSpan coercion into Etc
This provides a unified design where all coercion used by code for Axon functions will live in Etc.
We already have many coercion methods such as toId, toIds, toRec, toRecs, toGrid in Etc, so
it is a logical location for spans too.

But since spans are often passed as  function such as lastWeek, we should make things more
explicit that a context is required.  So adapt a pattern of accepting a HaystackContext in the
signature or if null then one is attempted to look up from actor local.

The span coercions require access to axon::ObjRange and axon::Fn to function.  So we move
the ObjRange class in Haystack (its a fairly generic tuple anyways).  And then add a new nodoc
mixin for HaystackFunc.  Lets allows all our logic to live in Etc cleanly.