hxConn: log warning when point has multiple fooConnRef tags
When configured this way, it creates a weird situation where both ConnLibs will map the point
a local ConnPoint in their lookup tables.  But the global ConnService will only see the last one.
So its indeterminate which one will win and be used with functions like connSyncHis, etc.
Then it gets more confusing if you remove one of the connRefs because there is no notification
to re-register it from another ConnLib.  To actually make it all work would require a lot of overhead,
so for now just log a warning which hopefully the admin will notice.