hxConn:  Fix boundary conditions to maintain conn watch state
Fix uncommon code paths which can lead to the ConnPoint.isWatched flag not getting set
correctly.  This can happen if a point is added to a connector which is already in watch.
But probably would also happen by adding haystackCur tag to an existing point in watch.

This fix includes:

- init isWatched flag inside ConnPoint constructor

- keep track of the ConnPoint.isWatched flag separately from isCurEnabled

- take care to call updatePointsInWatch on situations which might change points in watch