haystack: change Dict.each and eachWhile to iterate only non-null values
This is the first step in a series of changes to add static types into the haystack APIs using the xeto type system.
If we consider null a valid dict value, then it will cause endless problems when trying to apply a Maybe type.

This change only modifies the iterator signature.  It is ABI compatible, so does not require recompile.  However,
on the next recompile it might require fixes if an each/eachWhile block was checking for null.  It will also require a
source fix if anyone created their own Dict classes.

We have taken care in the Dict constructor methods (mostly on Etc) to continue to accept null without raising
an error as that would be a breaking change.

This change does have one potential breaking behavior with regards to JSON parsing.  Previously you could
serialize a dict with a null value to/from JSON.  But now null values will be stripped.