xeto: full support for nested specs in meta dict
We finally nail down how to deal with nested specs such as List<of:Ref<of:Spec>>.
In this design the spec meta for spec references is *always* a Ref with a qname value.
When parsing a spec that isn't a simple type reference, then we generate a synthetic
spec in the library using same underbar naming convention as unnamed specs.