xeto: remove Dict.spec method
This method requires that either a Dict pins a reference to a Spec or that the XetoEnv.cur
is used for lookup.  Both situations will cause massive future problems for environment
reload and multiple XetoEnvs active in one VM.  From this point onward, you must use Xeto.specOf
to resolve a Dict's spec Ref to an actual xeto::Spec instance now that we have standardized
the spec tag to be a Ref(qname).