xeto: enhance binary I/O to support inherited slots for an/or types
For simple single inheritance, its easy to optimize marshaling to just write
out the declared slots and compute inherited slots client side.  But the logic
for compound types like and/or is much more complicated - we don't want to
try to duplicate that logic with RSpec AST.  So instead just marshal the
inherited slot refs over the wire and use that to compute the effective slot map.