axon: pre-evaluate target arg in DotCall
This is prep work to dispatch on first arg for OO like Axon.
But first we need to eval the first arg safely, but only once and
still be able to dispatch to lazy funcs like readAll.  So eval, then
wrap as a Literal so its only evaluated once.

Note this change breaks the ability to use a dot call for functions
that expect a filter as the first parameter.  For example:

  readAll(temp)   // continues to work because not a dot call
  temp.readAll    // is now broken because temp is resolved as a variable