Updating publish.yaml file for github workflows, Dockerfile, and compose.yml file
Mostly finished with the three files, a lot is commented out from testing and edits
In the docker folder, the dev.Dockerfile is the main Dockerfile, it works with devDockerfile.compose.yml
Dockerfile can be ignored as of now, it will most likely be deleted later.
in .github/workflows publish-dev.Dockerfile.yaml is almost complete as well, it works with
dev.Dockerfile and devDockerfile.compose.yml to make the development build.