MQTT PubAck with Reason Code 16 (No Matching Subscribers) #58
Fixed decoding issue for PubAcks and a few other types. Was using
a const field on the type with the PacketType, but the decode()
constructor needed that information in order to properly decode.

The field is not initialized before the constructor completes and at the
time the field is needed by the decode, so I changed it to a method
instead and that resolves the issue.

Also fixed PacketWriterActor to trace the message before sending it
because sometimes the response was getting logged from the
PacketReaderActor before the writer had a chance to trace it.